[refactoring] extract events logic
